+++ e2e-cypress/files/plugins/index.js
@@ -16,8 +16,8 @@ module.exports = (on, config) => {
   // `config` is the resolved Cypress config
   on('task', {
     log(message) {
-      console.log(message)
-      return null
-    }
-  })
-}
+      console.log(message);
+      return null;
+    },
+  });
+};